performes double bypass surgery on the music player by gutting the soundcloud widget and replacing with the audio5js player. Removes the soundcloud-api.js file and adds the audio5js minified js file. Also moves the leaves animation js to a separate file in the js pipeline.
