+++ app/views/soundcloud/index.html.erb
@@ -1,7 +1,5 @@
 <div class="main">
 <div id="container" ng-app="SelmaApp">
-  <!-- SOUNDCLOUD IFRAME -->
-<!--   <iframe id="sc-widget" src=<%= "https://w.soundcloud.com/player/?url=https://api.soundcloud.com/tracks/188155459?client_id=#{ENV['SOUNDCLOUD_CLIENT_ID']}" %> width="300" height="300" scrolling="no" frameborder="no"></iframe> -->
 
   <div id="content-wrapper" ng-controller="patchController">
     <div id="leafContainer"></div>
@@ -30,9 +28,8 @@
     <!-- Section 2 - Player -->
     
     <article  id="scroll-link-2" class="scenarios">
-        <iframe id="sc-widget" src=<%= "https://w.soundcloud.com/player/?url=https://api.soundcloud.com/tracks/188155459?client_id=#{ENV['SOUNDCLOUD_CLIENT_ID']}" %> width="300" height="300" scrolling="no" frameborder="no"></iframe>
       <div class="music-player">
-        <i class="fa fa-play play-icon" ng-click="togglePlay()"></i>
+        <i class="fa fa-play play-icon"></i>
       </div>
 
         <!-- Question squares -->
@@ -101,47 +98,48 @@
       })
   });
 
-    var audioReady = function () {
+  var clickCount = -0.5;
+
+  var audioReady = function () {
+    // loads audio track from public directory
     this.load('/audios/music.mp3');
+    // maps trigger to play button
     var play_pause = document.querySelector('.play-icon');
     play_pause.addEventListener('click', playPause.bind(this));
+    var current_position;
+    // Binds playback counter to a constantly running number
+    this.on('timeupdate', function (position, duration) {
+      track_position = Math.floor(position);
+      if (track_position % 3 == 0 && track_position > 0 && track_position != current_position){
+        // Checks to make sure color squares do not load more than once for the same time position
+        current_position = track_position;
+        $('#questions').removeClass('hidden');
+        $('#questions p').addClass('expand');
+        $('.square').addClass('pop-up');
+      };
+    }, this);
   }
 
   var playPause = function () {
-    if (this.playing) {
-      this.pause();
-      this.volume(0);
-      console.log(this.position, this.duration, this.load_percent, this.volume());
-    } else {
-      this.play();
-      this.volume(1);
-    }
-    // or simply call this.playPause();
+    // sets to full volume
+    this.volume(1);
+    // toggles icon from play to pause
+    this.playPause();
+    $('.play-icon').toggleClass('fa-pause');
+      $('.play-icon').toggleClass('fa-play');
+    // console.log(this.position, this.duration, this.load_percent, this.volume());
   }
 
   var initAudio = function () {
+    // initiliazes audio player and sets backup default as audio5js.swf
     var audio5js = new Audio5js({
       swf_path: '/statics/swf/audio5js.swf',
       throw_errors: true,
-      format_time: true,
+      format_time: false,
       ready: audioReady
     });
   }
 
   initAudio();
-
-
-    // var audio5js = new Audio5js({
-    //   ready: function () {
-    //     this.load('/audios/music.mp3');
-    //     document.querySelector('.play-icon').addEventListener('click',
-    //       audio5js.playPause);
-    //   }
-    // });
-
-    //   document.querySelector('.play-icon').addEventListener('click', function () {
-    //    audio5js.playPause();
-    //    console.log("play toggled");
-    // }, false);
 
 </script>
